refactor(core): split COW implementation ownership
